Fold BindingMethodImplementation into MethodBindingExpression, and instead extract the decision making of whether+how to scope into a MethodImplementationStrategy.
Allowing BindingMethodImplementation to be included directly into MethodBindingExpression allows the implementation to make decisions on another axis: what type of method is being defined?

This allows us to correct the return type of private implementation methods in AOT. We do so by ignoring the check for public accessibility since the method will only be used in the same class.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=228766115
